Create AVFileScan.py
The code is a basic GUI program that allows users to select a file, scan it for viruses using VirusTotal API, and display the scan result in a label widget. The program also provides a link to the VirusTotal report page for the scanned file, which can be opened by clicking on the label. The user can choose a file by clicking on the "Select file" button, and initiate the scanning process by clicking on the "Scan file" button.